src/requirements.txt
openai-whisper
torch
torcheval
pytube
librosa
sentence-transformers
evaluate
numpy
soundfile

